bugfix in the dram test program

+++ drive/memory.c
@@ -1,5 +1,5 @@
 #include "memory.h"
 
-volatile uint32_t * get_ddr_base() {
-  return (uint32_t *)(DDR_RAM_BASE);
+volatile uint64_t * get_ddr_base() {
+  return (uint64_t *)(DDR_RAM_BASE);
 }
