+++ drive/memory.h
@@ -4,6 +4,6 @@
 #include <stdint.h>
 #include "device_map.h"
 
-extern volatile uint32_t * get_ddr_base();
+extern volatile uint64_t * get_ddr_base();
 
 #endif
